# docs/api/openapi/evolution-module-v1.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "SkillExecutionResult": {
        "additionalProperties": false,
        "properties": {
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "selection": {
            "$ref": "#/components/schemas/SkillSelection"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "selection",
          "score",
          "success"
        ],
        "title": "SkillExecutionResult",
        "type": "object"
      },
...
      "SkillSelection": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "company_id": {
            "maxLength": 48,
            "minLength": 1,
            "title": "Company Id",
            "type": "string"
          },
          "configuration": {
            "additionalProperties": true,
            "title": "Configuration",
            "type": "object"
          },
          "configuration_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Configuration Hash",
            "type": "string"
          },
          "deployment_id": {
            "title": "Deployment Id",
            "type": "string"
          },
          "experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "signature": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Signature",
            "type": "string"
          },
          "skill_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Skill Id",
            "type": "string"
          },
          "trace_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "company_id",
          "agent_id",
          "skill_id",
          "trace_id",
          "deployment_id",
          "experiment_id",
          "version",
          "configuration_hash",
          "configuration",
          "expires_at",
          "signature"
        ],
        "title": "SkillSelection",
        "type": "object"
      },
      "SkillSelectionRequest": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "company_id": {
            "maxLength": 48,
            "minLength": 1,
            "title": "Company Id",
            "type": "string"
          },
          "skill_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Skill Id",
            "type": "string"
          },
          "trace_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "company_id",
          "agent_id",
          "skill_id",
          "trace_id"
        ],
        "title": "SkillSelectionRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Evolution capability for global trace analysis and architecture recommendations.",
    "title": "Evolution Module",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evolution/skill-executions/resolve": {
      "post": {
        "operationId": "resolve_api_v1_evolution_skill_executions_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillSelection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve"
      }
    },
    "/api/v1/evolution/skill-executions/results": {
      "post": {
        "operationId": "record_api_v1_evolution_skill_executions_results_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillExecutionResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Record Api V1 Evolution Skill Executions Results Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record"
      }
    },
...
  }
}